Finalize main.js
All issues have been addressed except for one: customize whitelist. A description is given near the end of main.js of what that function would entail, noting that it is yet to be implemented, and gives a justification for why it has yet to be implemented.

Updated commenting and separated code into more functions for the sake of cohesion and organization.